big_sister/requirements.txt
PyTorch
nltk
numpy